@@ -1,7 +1,7 @@
 package io.github.ryanhoo.music.ui.base;
 
 import android.graphics.Point;
-import android.support.v4.app.DialogFragment;
+import androidx.fragment.app.DialogFragment;
 import android.view.Window;
 import android.view.WindowManager;
 
