@@ -2,8 +2,8 @@ package io.github.ryanhoo.music.ui.base;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 import rx.Subscription;
 import rx.subscriptions.CompositeSubscription;
