@@ -1,7 +1,7 @@
 package io.github.ryanhoo.music.ui.common;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
