@@ -1,7 +1,7 @@
 package io.github.ryanhoo.music.ui.local.filesystem;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
 import io.github.ryanhoo.music.ui.base.adapter.ListAdapter;
