@@ -1,9 +1,9 @@
 package io.github.ryanhoo.music.ui.playlist;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.widget.PopupMenu;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.Nullable;
+import androidx.appcompat.widget.PopupMenu;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.*;
 import android.widget.ProgressBar;
 import android.widget.Toast;
