@@ -5,8 +5,8 @@ import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.os.Build;
-import android.support.annotation.ColorInt;
-import android.support.annotation.Dimension;
+import androidx.annotation.ColorInt;
+import androidx.annotation.Dimension;
 import android.util.AttributeSet;
 import android.view.View;
 import io.github.ryanhoo.music.R;
